Update hook_views_data_alter() to fix issue #789
I think there are still some adjustments to be made: some $data[<table>]['table']['join'] may not be needed and I did not succeed to add the relationship from a Tripal entity  to a Chado object. However, from Chado to Tripal entity, it is working. At least, it's better than current code that does nothing!