Changed conditional for optional treatment (WIP)
Optional can now trigger on non-required records. This has unfortunately produced a bug that seems to be linked to inaccurate record tracking. Bulk loader is only able to iterate one row of an input file at a time (crashes on the second newest row). Processed rows appear to be accurate.