Finished tripal field item base.
Finished adding all necessary methods to the tripal field item base class. This
includes the schema and property definitions that are auto generated using the
tripal field item base interface. This class was also moved to a base class
namespace for tripal fields because it is the base class and not an actual
implementation of a tripal field.