Fixed field plugin bugs.
Fixed multiple bugs related to including other classes in separate files.

Fixed class declaration bugs pertaining to invalid syntax.

Moved tripal field item interface and tripal storage interface files to new
folders and removed the interface folder. Naming a folder interface is not
compatible with symfony/PHP because it is a keyword in the PHP language.