Fix install issue when Tripal is installed by a distribution (profile)
When Tripal is installed by a distribution (profile), no database parameters are set in settings.php yet, resulting in a "The specified database connection is not defined: default" error because of "\Drupal::database()->getConnectionOptions()" being called. To avoid that issue, a try-catch statement has been added to return an empty array when the $databases variable is not set.