Addressed the problem that there was a discrepancy between bulk publishing by CahdoPublish and editing an entry via the UI. Not all the fields from ancilliary table would be filled or updated by editing the entity. Now, the same logic as during bulk entity publishing is applied.
To accomplish this, two additions were made:

 - TripalEntity now overrides postSave. This method invokes the new update method in ChadoPublish
 - The method updatePublishedEntity has been added to ChadoPublish to fill all fields in the same way as publish does

The updatePublishedEntity method in the ChadoPublish class is responsible for
updating a specific field of a Tripal entity in a Drupal application.
The method uses the same initialization and field value retrieval as publish.

An SQL update query is constructed using the Drupal connection object, setting the fields
to be updated and adding conditions for the entity_id and delta.
The query is executed within a try-catch block to gracefully handle any exceptions.