First minimal WORKING example.
To test, import the PAU8 gene GFF and compare residues with the Sequence from Alignment.

They should be identical, except for non-capital base codes from the genome.

The id of the current record is now added to the context array.
The id is used in the sequence helper to retrieve the gene sequence from alignment.

A few warnings from the retrieval code.